requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dominate==2.3.5
falcon==2.0.0
Flask==1.0.3
Flask-Bootstrap4==4.0.2
Flask-JWT==0.3.2
idna==2.8
importlib-metadata==0.18
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
pkg-resources==0.0.0
pluggy==0.12.0
py==1.8.0
PyJWT==1.4.2
pylint==2.3.1
pyparsing==2.4.0
pytest==4.6.3
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.2
zipp==0.5.1
